requirements.txt
pika==1.2.0
google-cloud-storage==2.5.0